fix: Align with Kùzu: change the return type of length() to int64_t (#680)
This pull request updates the handling of path lengths in the graph
database engine to use 64-bit integers instead of 32-bit. This change
ensures that the system can support much longer paths without risk of
integer overflow.

**Path length type changes:**

* Changed the `elem_t` type in `PathLenPathAccessor` from `int32_t` to
`int64_t` to allow for larger path lengths.
* Updated the return type of `eval_path` in `PathLenPathAccessor` to use
`int64_t` instead of `int32_t`.
* Changed the return type of the `len()` method in the `Path` class from
`int32_t` to `int64_t` for consistency and to support longer paths.

+++ src/neug/engines/graph_db/runtime/common/accessors.h
@@ -691,7 +691,7 @@ class PathIdPathAccessor : public IAccessor {
 
 class PathLenPathAccessor : public IAccessor {
  public:
-  using elem_t = int32_t;
+  using elem_t = int64_t;
   PathLenPathAccessor(const Context& ctx, int tag)
       : path_col_(*std::dynamic_pointer_cast<IPathColumn>(ctx.get(tag))) {}
 
@@ -700,7 +700,7 @@ class PathLenPathAccessor : public IAccessor {
   }
 
   RTAny eval_path(size_t idx) const override {
-    return RTAny(static_cast<int32_t>(typed_eval_path(idx)));
+    return RTAny(static_cast<int64_t>(typed_eval_path(idx)));
   }
 
  private:
